requirements.txt
Flask
Flask-WTF
Flask-Babel
passlib
python-ldap
ldappool
wrapt